Divide all classes in different modules to handle better the whole app
