+++ README.md
@@ -5,12 +5,13 @@ This repository contains an web app that allows to add and remove items to a to-
 ## Built With
 
 - HTML
-- CSS
+- SCSS
 - Javascript
 - Git & Github
 - WebPack
 
 ## Live Demo
+
 - Check the live demo of my page [here](https://ulisesedu7.github.io/To-Do-List-Project/)
 
 ## Getting Started
